+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Display object depth in flash panel
+- Show imported files on script import, able to cancel import
 
 ### Fixed
 - [#1761] AS3 - try..finally inside another structure like if